Now able to read in audio data

+++ src/main/java/com/dakkra/hypersynesthesia/HSAudioBuffer.java
@@ -5,19 +5,19 @@ import java.util.Vector;
 
 public class HSAudioBuffer {
 
-	private List<Short> buffer;
+	private List<Integer> buffer;
 
 	public HSAudioBuffer() {
 		this.buffer = new Vector<>();
 	}
 
-	public void addSample( short sample ) {
+	public void addSample( int sample ) {
 		buffer.add( sample );
 	}
 
-	public void addSamples( short[] samples ) {
+	public void addSamples( int[] samples ) {
 		if( samples == null ) throw new NullPointerException( "Samples cannot be null" );
-		for( short sample : samples ) {
+		for( int sample : samples ) {
 			addSample( sample );
 		}
 	}
