+++ src/main/java/com/dakkra/hypersynesthesia/HSStereoAudioBuffer.java
@@ -11,12 +11,12 @@ public class HSStereoAudioBuffer {
 		this.rightBuffer = new HSAudioBuffer();
 	}
 
-	public void addSample( short leftSample, short rightSample ) {
+	public void addSample( int leftSample, int rightSample ) {
 		leftBuffer.addSample( leftSample );
 		rightBuffer.addSample( rightSample );
 	}
 
-	public void addSamples( short[] leftSamples, short[] rightSamples ) {
+	public void addSamples( int[] leftSamples, int[] rightSamples ) {
 		if( leftSamples == null || rightSamples == null ) throw new NullPointerException( "Samples cannot be null" );
 		if( leftSamples.length != rightSamples.length ) throw new IllegalArgumentException( "Left and right samples must be the same length" );
 		for( int i = 0; i < leftSamples.length; i++ ) {
